Add confirm delete

+++ client/src/components/jobs/JobItem.js
@@ -1,27 +1,33 @@
-import React from 'react'
+import React from "react";
 import Moment from "react-moment";
-import { connect } from 'react-redux'
-import {deleteJob} from '../../actions/job'
-import Avatar from 'react-avatar';
+import { connect } from "react-redux";
+import { deleteJob } from "../../actions/job";
+import Avatar from "react-avatar";
 
-const JobItem = ({auth, job, deleteJob}) => {
+const JobItem = ({ auth, job, deleteJob }) => {
 	const onDelete = () => {
-		deleteJob(job._id)
-	}
+		if (window.confirm("Delete?")) {
+			deleteJob(job._id);
+		}
+	};
 	return (
 		<li className="collection-item">
 			<div>
 				<span
-					className='title' 					
+					className="title"
 					// onClick={() => setCurrent(log)}
 				>
-					<Avatar name={job.author} size="60" round={true} /> {' '}
+					<Avatar name={job.author} size="60" round={true} />{" "}
 					{job.author}
-					<span className="secondary-content">{job.industry}, {job.employment}</span>
+					<span className="secondary-content">
+						{job.industry}, {job.employment}
+					</span>
 				</span>
 				<br />
-				<span className='title'>
-				<h5>{job.title} at {job.company}</h5>
+				<span className="title">
+					<h5>
+						{job.title} at {job.company}
+					</h5>
 				</span>
 				<br />
 				<span className="title">{job.description}</span>
@@ -40,17 +46,21 @@ const JobItem = ({auth, job, deleteJob}) => {
 						{job.date}
 					</Moment>
 					{!auth.loading && job.user === auth.user._id && (
-					<button onClick={onDelete} className="secondary-content">
-					<i className="material-icons grey-text">delete</i>
-					</button>)}
+						<button
+							onClick={onDelete}
+							className="secondary-content"
+						>
+							<i className="material-icons grey-text">delete</i>
+						</button>
+					)}
 				</span>
 			</div>
 		</li>
-	)
-}
+	);
+};
 
 const mapStateToProps = ({ auth }) => {
 	return { auth };
 };
 
-export default connect(mapStateToProps, {deleteJob})(JobItem)
+export default connect(mapStateToProps, { deleteJob })(JobItem);
